Removed warnings due to use of int64_t in several files. Using lld() mask to print int64_t numbers.
SVN revision 7221

+++ src/proto/PINT-le-bytefield.c
@@ -586,8 +586,9 @@ static int lebf_decode_req(
 
     if (ptr != (char *) input_buffer + input_size)
     {
-	gossip_lerr("%s: op %d consumed %d bytes, but message was %d bytes.\n",
-                    __func__, req->op, ptr - (char *) input_buffer, input_size);
+	gossip_lerr("%s: op %u consumed %ld bytes, but message was %d bytes.\n",
+                    __func__, req->op, 
+				(unsigned long)(ptr - (char *) input_buffer), input_size);
 	ret = -PVFS_EPROTO;
     }
 
@@ -675,8 +676,9 @@ static int lebf_decode_resp(
 #undef CASE
 
     if (ptr != (char *) input_buffer + input_size) {
-	gossip_lerr("%s: op %d consumed %d bytes, but message was %d bytes.\n",
-                    __func__, resp->op, ptr - (char *) input_buffer,
+	gossip_lerr("%s: op %u consumed %ld bytes, but message was %d bytes.\n",
+                    __func__, resp->op, 
+				(unsigned long)(ptr - (char *) input_buffer),
                     input_size);
 	ret = -PVFS_EPROTO;
     }
